Made the process part of the ADBView.
Makes it possible to have multiple adb logcat processes running,
with the buffer name being the name of the device.
Also added a setting for whether to just launch adb or not when
there's only a single device connected.
Also added the current adb views to the ADB launch list, which will
be focused if selected.
Made the default adb_args include timestamps.

Closes #5.